+++ input_letter.h
@@ -13,6 +13,6 @@ void print_input_letter(const InputLetter *const list, int len);
 void letter_set_status(InputLetter *list, int i, char c);
 
 // get the accuracy
-int letter_accuracy(const InputLetter *const list, int len);
+double letter_accuracy(const InputLetter *const list, int len);
 
 #endif
